refactor: reorganize authentication components and update import paths
